@@ -17,18 +17,20 @@
     "fix": "prettier --write ."
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.3",
-    "@docusaurus/plugin-content-docs": "^2.4.3",
-    "@docusaurus/theme-classic": "^2.4.3",
-    "@mdx-js/react": "^1.6.22",
-    "react": "^17.0.2"
+    "@docusaurus/core": "^3.0.1",
+    "@docusaurus/plugin-content-docs": "^3.0.1",
+    "@docusaurus/theme-classic": "^3.0.1",
+    "@mdx-js/react": "^3.0.0",
+    "react": "^18.2.0",
+    "remark-wiki-link": "^2.0.1"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.3",
-    "@docusaurus/theme-common": "^2.4.3",
-    "@docusaurus/types": "^2.4.3",
-    "@tsconfig/docusaurus": "^2.0.1",
-    "@types/node": "^20.7.1",
+    "@docusaurus/module-type-aliases": "^3.0.1",
+    "@docusaurus/theme-common": "^3.0.1",
+    "@docusaurus/tsconfig": "^3.0.1",
+    "@docusaurus/types": "^3.0.1",
+    "@types/node": "^20.8.10",
+    "@types/react": "^18.2.36",
     "prettier": "^3.0.3",
     "typescript": "5.2.2"
   },
